Moving init socketServer to app.js
