+++ app.js
@@ -1,48 +1,48 @@
-const log = require('simple-node-logger').createSimpleLogger();
+const log = require("simple-node-logger").createSimpleLogger();
 
-const http = require('http');
+const http = require("http");
 
-const envProvider = require('./env_provider');
+const envProvider = require("./env_provider");
 
-const express = require('express');
+const express = require("express");
 
 const app = express();
 
 const port = envProvider.SRV_PORT;
 
-app.set('port', port);
+app.set("port", port);
 
 const server = http.createServer(app);
 
 // TODO: Delete this, no need to dispatch routes anymore for this app
 // require('./routes/router').routeDispatcher(app);
-const Organizer = require('./socketServer').Organizer;
+const Organizer = require("./socketServer").Organizer;
 organizer = new Organizer(server);
 
 server.listen(port);
 
-server.on('error', onError);
-server.on('listening', onListening);
+server.on("error", onError);
+server.on("listening", onListening);
 
 /**
  * Event listener for HTTP server "error" event.
  * @param {Object} error returned.
  */
 function onError(error) {
-  if (error.syscall !== 'listen') {
+  if (error.syscall !== "listen") {
     throw error;
   }
 
-  const bind = typeof port === 'string' ? 'Pipe ' + port : 'Port ' + port;
+  const bind = typeof port === "string" ? "Pipe " + port : "Port " + port;
 
   // handle specific listen errors with friendly messages
   switch (error.code) {
-    case 'EACCES':
-      log.error(bind + ' requires elevated privileges');
+    case "EACCES":
+      log.error(bind + " requires elevated privileges");
       process.exit(1);
       break;
-    case 'EADDRINUSE':
-      log.error(bind + ' is already in use');
+    case "EADDRINUSE":
+      log.error(bind + " is already in use");
       process.exit(1);
       break;
     default:
@@ -55,9 +55,9 @@ function onError(error) {
  */
 function onListening() {
   const addr = server.address();
-  const bind = typeof addr === 'string' ? 'pipe ' + addr : 'port ' + addr.port;
+  const bind = typeof addr === "string" ? "pipe " + addr : "port " + addr.port;
   log.info(bind);
-  log.info('server started successfully, listening to : ' + addr.port);
+  log.info("server started successfully, listening to : " + addr.port);
 }
 
 module.exports = app;
